Fix some build issues and warnings
Removed unused function which failed to link on Linux,
fixed some signedness warnings.

+++ QueryRenderer/Color.h
@@ -62,7 +62,7 @@ class ColorRGBA {
   operator std::string() const {
     std::ostringstream s("[");
 
-    for (int i = 0; i < _colorArray.size(); ++i) {
+    for (size_t i = 0; i < _colorArray.size(); ++i) {
       if (i > 0) {
         s << ", ";
       }
